+++ setup.py
@@ -10,9 +10,9 @@ setup( name='gratipay'
                       : [ 'payday=gratipay.cli:payday'
                         , 'fake_data=gratipay.utils.fake_data:main'
                         , 'sync-npm=gratipay.sync_npm.cli:main'
-                        , 'queue-branch-email=gratipay.utils.emails.queue_branch_email:main'
-                        , 'dequeue-emails=gratipay.utils.emails.dequeue_emails:main'
-                        , 'list-email-queue=gratipay.utils.emails.list_email_queue:main'
+                        , 'queue-branch-email=gratipay.email.cli.queue_branch_email:main'
+                        , 'dequeue-emails=gratipay.email.cli.dequeue_emails:main'
+                        , 'list-email-queue=gratipay.email.cli.list_email_queue:main'
                          ]
                        }
       )
