+++ www/~/%username/emails/verify.html.spt
@@ -4,7 +4,8 @@ from datetime import timedelta
 
 from aspen import Response
 from aspen.utils import utcnow
-from gratipay.utils import emails, get_participant, decode_from_querystring
+from gratipay.utils import get_participant, decode_from_querystring
+from gratipay.models.participant import email
 
 [-----------------------------------------------------------------------------]
 
@@ -17,11 +18,11 @@ if participant == user.participant:
         # and this can be removed when we're confident that there are no more
         # old-style verification links hanging out there.
 
-        email = request.qs['email']
+        email_address = request.qs['email']
     else:
-        email = decode_from_querystring(request.qs.get('email2', ''), default='')
+        email_address = decode_from_querystring(request.qs.get('email2', ''), default='')
     nonce = request.qs.get('nonce', '')
-    result = participant.verify_email(email, nonce)
+    result = participant.verify_email(email_address, nonce)
     if not participant.email_lang:
         participant.set_email_lang(request.headers.get("Accept-Language"))
 
@@ -55,33 +56,33 @@ suppress_sidebar = True
                 '<a href="mailto:support@gratipay.com">support@gratipay.com</a>'|safe) }}</p>
 
     {% else %}
-        {% if result == emails.VERIFICATION_SUCCEEDED %}
+        {% if result == email.VERIFICATION_SUCCEEDED %}
             <h1>{{ _("Success!") }}</h1>
             <p>{{ _("Your email address {0} is now connected to your Gratipay account.",
                     ("<b>%s</b>"|safe) % email) }}</p>
-        {% elif result == emails.VERIFICATION_STYMIED %}
+        {% elif result == email.VERIFICATION_STYMIED %}
             <h1>{{ _("Address Taken") }}</h1>
             <p>{{ _("The email address {0} is already connected to a different Gratipay account.",
-                    ("<b>%s</b>"|safe) % email) }}</p>
-        {% elif result == emails.VERIFICATION_REDUNDANT %}
+                    ("<b>%s</b>"|safe) % email_address) }}</p>
+        {% elif result == email.VERIFICATION_REDUNDANT %}
             <h1>{{ _("Already Verified") }}</h1>
             <p>{{ _("Your email address {0} is already connected to your Gratipay account.",
-                    ("<b>%s</b>"|safe) % email) }}</p>
-        {% elif result == emails.VERIFICATION_MISSING %}
+                    ("<b>%s</b>"|safe) % email_address) }}</p>
+        {% elif result == email.VERIFICATION_MISSING %}
             <h1>{{ _("Missing Info") }}</h1>
             <p>{{ _("Sorry, that's a bad link. You'll need to view your email addresses "
                     "and start over.") }}</p>
         {% else %}
-            {% if result == emails.VERIFICATION_EXPIRED  %}
+            {% if result == email.VERIFICATION_EXPIRED  %}
                 <h1>{{ _("Expired") }}</h1>
                 <p>{{ _("The verification code for {0} has expired.",
-                        ("<b>%s</b>"|safe) % email) }}</p>
-            {% elif result == emails.VERIFICATION_FAILED %}
+                        ("<b>%s</b>"|safe) % email_address) }}</p>
+            {% elif result == email.VERIFICATION_FAILED %}
                 <h1>{{ _("Failure") }}</h1>
                 <p>{{ _("The verification code for {0} is bad.",
-                        ("<b>%s</b>"|safe) % email) }}</p>
+                        ("<b>%s</b>"|safe) % email_address) }}</p>
             {% endif %}
-            <p data-email="{{ email }}"><button class="resend">{{ _("Resend verification email") }}</button></p>
+            <p data-email="{{ email_address }}"><button class="resend">{{ _("Resend verification email") }}</button></p>
         {% endif %}
         <a href="/{{ participant.username }}/emails/">{{ _("View your email addresses") }}.</a>
     {% endif %}
